use MEF to call InfoBar
