+++ src/NuGet.Clients/NuGet.SolutionRestoreManager/IInfoBarService.cs
@@ -1,0 +1,13 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace NuGet.SolutionRestoreManager
+{
+    public interface IInfoBarService
+    {
+        Task ShowInfoBar(CancellationToken cancellationToken);
+    }
+}
