+++ src/NuGet.Clients/NuGet.Tools/GlobalSuppressions.cs
@@ -28,3 +28,4 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Build", "CA1031:Modify 'WriteUserOptions' to catch a more specific allowed exception type, or rethrow the exception.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.PackageManagement.UI.SolutionUserOptions.WriteUserOptions(Microsoft.VisualStudio.OLE.Interop.IStream,System.String)~System.Int32")]
 [assembly: SuppressMessage("Usage", "CA1816:Dispose methods should call SuppressFinalize", Justification = "IVsPackage.Close() method invokes overridden NuGetPackage.Dispose() method which internally invokes GC.SuppressFinalize() method.", Scope = "member", Target = "~M:NuGetVSExtension.NuGetPackage.Dispose(System.Boolean)")]
 [assembly: SuppressMessage("Design", "CA1001:Types that own disposable fields should be disposable", Justification = "IVsPackage.Close() method invokes overridden NuGetPackage.Dispose() method upon closing the IDE.", Scope = "type", Target = "~T:NuGetVSExtension.NuGetPackage")]
+[assembly: SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVSExtension.PMUIStarterDelegateService.CreateDocWindowForSolutionAsync~System.Threading.Tasks.Task{Microsoft.VisualStudio.Shell.Interop.IVsWindowFrame}")]
