+++ src/NuGet.Clients/NuGet.VisualStudio.Common/Services/IPMUIStarter.cs
@@ -1,0 +1,10 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.VisualStudio
+{
+    public interface IPMUIStarter
+    {
+        void PMUIStarter();
+    }
+}
